added support for unDIMed variables - all subscripts found during variable setup are now set to a minimum of 11 - new cases in tests to make sure this works - updated TODO list to remove it, but also added long-variable-names, which we should do
NOTE: because the variable setup takes place during the parsing, the calls into insert_variable do not know if they are inside a DIM or not. This means that DIM A(3) will still set aside 11 slots, whereas in MS it would be 0..3. This means it will not return an error on A(4), which MS does. This is not worth fixing.